Remove the old api plugin.
